+++ ballerina-tests/Dependencies.toml
@@ -11,7 +11,6 @@ distribution-version = "2201.11.0"
 org = "ballerina"
 name = "auth"
 version = "2.13.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "crypto"},
 	{org = "ballerina", name = "jballerina.java"},
@@ -24,7 +23,6 @@ dependencies = [
 org = "ballerina"
 name = "cache"
 version = "3.9.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "constraint"},
 	{org = "ballerina", name = "jballerina.java"},
@@ -36,7 +34,6 @@ dependencies = [
 org = "ballerina"
 name = "constraint"
 version = "1.6.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"}
 ]
@@ -45,7 +42,6 @@ dependencies = [
 org = "ballerina"
 name = "crypto"
 version = "2.8.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"},
 	{org = "ballerina", name = "time"}
@@ -55,7 +51,6 @@ dependencies = [
 org = "ballerina"
 name = "data.jsondata"
 version = "1.0.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"},
 	{org = "ballerina", name = "lang.object"}
@@ -65,7 +60,6 @@ dependencies = [
 org = "ballerina"
 name = "file"
 version = "1.11.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "io"},
 	{org = "ballerina", name = "jballerina.java"},
@@ -77,7 +71,6 @@ dependencies = [
 org = "ballerina"
 name = "http"
 version = "2.13.4"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "auth"},
 	{org = "ballerina", name = "cache"},
@@ -107,7 +100,6 @@ dependencies = [
 org = "ballerina"
 name = "io"
 version = "1.7.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"},
 	{org = "ballerina", name = "lang.value"}
@@ -125,7 +117,6 @@ version = "0.0.0"
 org = "ballerina"
 name = "jwt"
 version = "2.14.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "cache"},
 	{org = "ballerina", name = "crypto"},
@@ -141,7 +132,6 @@ dependencies = [
 org = "ballerina"
 name = "lang.__internal"
 version = "0.0.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"},
 	{org = "ballerina", name = "lang.object"}
@@ -151,7 +141,6 @@ dependencies = [
 org = "ballerina"
 name = "lang.array"
 version = "0.0.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"},
 	{org = "ballerina", name = "lang.__internal"}
@@ -161,7 +150,6 @@ dependencies = [
 org = "ballerina"
 name = "lang.decimal"
 version = "0.0.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"}
 ]
@@ -179,7 +167,6 @@ dependencies = [
 org = "ballerina"
 name = "lang.int"
 version = "0.0.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"},
 	{org = "ballerina", name = "lang.__internal"},
@@ -190,13 +177,11 @@ dependencies = [
 org = "ballerina"
 name = "lang.object"
 version = "0.0.0"
-scope = "testOnly"
 
 [[package]]
 org = "ballerina"
 name = "lang.regexp"
 version = "0.0.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"}
 ]
@@ -208,7 +193,6 @@ modules = [
 org = "ballerina"
 name = "lang.runtime"
 version = "0.0.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"}
 ]
@@ -217,7 +201,6 @@ dependencies = [
 org = "ballerina"
 name = "lang.string"
 version = "0.0.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"},
 	{org = "ballerina", name = "lang.regexp"}
@@ -227,7 +210,6 @@ dependencies = [
 org = "ballerina"
 name = "lang.value"
 version = "0.0.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"}
 ]
@@ -236,7 +218,6 @@ dependencies = [
 org = "ballerina"
 name = "log"
 version = "2.11.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "io"},
 	{org = "ballerina", name = "jballerina.java"},
@@ -248,7 +229,6 @@ dependencies = [
 org = "ballerina"
 name = "mime"
 version = "2.11.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "io"},
 	{org = "ballerina", name = "jballerina.java"},
@@ -260,7 +240,6 @@ dependencies = [
 org = "ballerina"
 name = "oauth2"
 version = "2.13.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "cache"},
 	{org = "ballerina", name = "crypto"},
@@ -282,7 +261,6 @@ dependencies = [
 org = "ballerina"
 name = "os"
 version = "1.9.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "io"},
 	{org = "ballerina", name = "jballerina.java"}
@@ -292,7 +270,6 @@ dependencies = [
 org = "ballerina"
 name = "task"
 version = "2.6.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"},
 	{org = "ballerina", name = "time"}
@@ -316,7 +293,6 @@ modules = [
 org = "ballerina"
 name = "time"
 version = "2.6.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"}
 ]
@@ -325,7 +301,6 @@ dependencies = [
 org = "ballerina"
 name = "url"
 version = "2.5.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"}
 ]
@@ -334,7 +309,6 @@ dependencies = [
 org = "ballerina"
 name = "xmldata"
 version = "2.9.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "jballerina.java"}
 ]
@@ -343,7 +317,6 @@ dependencies = [
 org = "ballerina"
 name = "yaml"
 version = "0.7.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "file"},
 	{org = "ballerina", name = "io"},
@@ -368,7 +341,6 @@ modules = [
 org = "ballerinax"
 name = "ai.agent"
 version = "0.7.9"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "file"},
 	{org = "ballerina", name = "http"},
@@ -406,7 +378,6 @@ modules = [
 org = "ballerinax"
 name = "azure.openai.chat"
 version = "3.0.1"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "constraint"},
 	{org = "ballerina", name = "http"},
@@ -418,7 +389,6 @@ dependencies = [
 org = "ballerinax"
 name = "openai.chat"
 version = "3.0.0"
-scope = "testOnly"
 dependencies = [
 	{org = "ballerina", name = "constraint"},
 	{org = "ballerina", name = "http"},
