+++ ballerina/types.bal
@@ -14,12 +14,14 @@
 // specific language governing permissions and limitations
 // under the License.
 
+import ballerina/http;
+
 # Represents a request message for the chat service.
 #
-# + session_id - A unique identifier for the chat session.
+# + sessionId - A unique identifier for the chat session.
 # + message - The content of the chat message sent by the user.
 public type ChatReqMessage record {|
-    string session_id;
+    string sessionId;
     string message;
 |};
 
@@ -28,6 +30,11 @@ public type ChatReqMessage record {|
 # + message - The response message generated by the chat service.
 public type ChatRespMessage record {|
     string message;
+|};
+
+# Represents the configuration for a chat client.
+public type ChatClientConfiguration record {|
+    *http:ClientConfiguration;
 |};
 
 # Defines a chat service interface that handles incoming chat messages.
